Refactored code and added firebase
